% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAttribute.R
\name{getAttribute}
\alias{getAttribute}
\title{Get data from CCM for a specified case}
\usage{
getAttribute(caseId, optionsList)
}
\arguments{
\item{caseId}{character scalar. Names the case to
obtain information about.}

\item{optionsList}{list object. Controls the behaviour
of the query, including the filter for the query, and
data to return.}
}
\value{
If the query succeeds, a \code{list()} containing data
for the specified \code{caseId}. Otherwise, a \code{list()} of
\code{NA} values.
}
\description{
\code{getAttribute()} queries CCM for details about
a case. the parameter \code{optionsList} specifies the
columns to return.
}
