% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDBLabels.R
\name{getDBLabels}
\alias{getDBLabels}
\title{Helper function for building SOQL queries}
\usage{
getDBLabels(table, colNames)
}
\arguments{
\item{table}{Character scalar. Names the CCM object
which the function will use to search for field names.}

\item{colNames}{Character vector. Contains a combination
field names and labels to search for in \code{table}.}
}
\value{
Character vector. Field names suitable for use in
subsequent SOQL queries. Elements form \code{colNames} not found
in \code{table} are removed from the results.
}
\description{
\code{getDBLabels()} retrieves the description of a specified
object in CCM then searches for field names using the
field label seen on the client-side.
}
