% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getExposureAttribute.R
\name{getExposureAttribute}
\alias{getExposureAttribute}
\title{Get attributes related to an exposure}
\usage{
getExposureAttribute(attribute, exposure)
}
\arguments{
\item{attribute}{character scalar. Names the CCM object used to
obtain information about an exposure. One of the currently
supported objects:
\enumerate{
\item \code{investigations}: Exposure Investigations
}}

\item{exposure}{character scalar or vector. Names the CCM Exposure Id
to obtain information about.}
}
\value{
a \code{tibble} of information related to an exposure.
}
\description{
\code{getExposureAttribute()} specifies the object used to
get related information about an exposure. This function
wraps \code{getAttribute()} to allow iteration over multiple
exposures. See \code{getAttribute()}.
}
\seealso{
\code{\link[=getAttribute]{getAttribute()}} for information on how the CCM
query is executed. \code{\link[=getExposures]{getExposures()}} for obtaining Exposure Id's
required for this function.
}
