% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login.R
\name{login}
\alias{login}
\title{Obtain an access token to authroize requests}
\usage{
login()
}
\value{
CCM access token for use in authorizing subsequent requests.
}
\description{
This function implements the OAuth2.0 device grant flow
see \url{https://tools.ietf.org/html/rfc6749#section-1.3.1}
and \url{https://help.salesforce.com/articleView?id=remoteaccess_oauth_device_flow.htm&type=5}
for more information. The token is stored in the OS keyring.
}
