% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAttribute.R
\name{getAttribute}
\alias{getAttribute}
\title{Get data from CCM for a specified record}
\usage{
getAttribute(Id, optionsList)
}
\arguments{
\item{Id}{Character scalar. Names the record to
obtain information about.}

\item{optionsList}{list object. Controls the behaviour
of the query including the filter for the query, and
data to return.}
}
\value{
If the query succeeds, a \code{tibble} containing data
for the specified \code{Id}. Otherwise, a \code{tibble} of
\code{NA} values.
}
\description{
\code{getAttribute()} queries CCM for details about
a record. the parameter \code{optionsList} specifies the
columns to return.
}
