% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeTimestamp.R
\name{makeTimestamp}
\alias{makeTimestamp}
\title{Helper function for building ISO-8601 dates.}
\usage{
makeTimestamp(dateString)
}
\arguments{
\item{dateString}{Character scalar. Names the date
which the function will use to create the formatted
datetime.}
}
\value{
Character scalar. An ISO-8601 formatted date
in the UTC timezone suitable for use in CCM queries.
}
\description{
\code{makeTimestame()} generates ISO-8601 formatted
date strings in UTC. This is the format required
to query datetime fields in CCM.
}
